local illumination shading project
This was my first programming assignment in CS-184: Foundations of
Computer Graphics with Professor James O'Brien. It is a
local-illumination demo using the Phong reflection model. I wrote it in
Cocoa on Mac OS X because of Cocoa's intuitive user-interface design.

This particular implementation has support for drawing spheres, planes,
and tori, using either Phong, "toon", or sub-surface scattering shaders.